Addressed Raz's comments, moved get_pv_for_pvc() to pvc.py as static
