+++ tests/test_fio_workload.py
@@ -6,7 +6,7 @@ from ocs_ci.ocs import workload
 from ocs_ci.framework.testlib import ManageTest
 from ocs_ci.utility.utils import TimeoutSampler
 from ocs_ci.utility import templating
-from tests import helpers
+from ocs_ci.ocs.resources import pvc
 
 
 from tests.fixtures import (
@@ -28,18 +28,13 @@ def test_fixture(request):
     def finalizer():
         teardown(self)
     request.addfinalizer(finalizer)
-    setup(self)
-
-
-def setup(self):
-    pass
 
 
 def teardown(self):
     logger.info(f"Deleting pod {self.pod_obj.name}")
     self.pod_obj.delete()
     self.pvc_obj.reload()
-    pv = helpers.get_pv_for_pvc(self.pvc_obj)
+    pv = pvc.get_pv_for_pvc(self.pvc_obj)
     logger.info(f"PV for PVC is {pv.name}")
     logger.info(f"Deleting PVC {self.pvc_obj.name}")
     self.pvc_obj.delete()
